Revert and manually applied "optimised chat rendering a bit" (commit 0321c97070d5d4fabd79c91aee72b4136cf93306)

+++ src/game/editor/ed_editor.cpp
@@ -1894,7 +1894,7 @@ static void ExtractName(const char *pFileName, char *pName, int BufferSize)
 			pEnd = pFileName;
 	}
 
-	int Length = pEnd > pExtractedName ? min(BufferSize, pEnd-pExtractedName+1) : BufferSize;
+	int Length = pEnd > pExtractedName ? min(BufferSize, (int)(pEnd-pExtractedName+1)) : BufferSize;
 	str_copy(pName, pExtractedName, Length);
 }
